+++ features/profile/components/atoms/follow-button.tsx
@@ -1,10 +1,9 @@
 'use client';
 
-import { useMutation, useQueryClient } from '@tanstack/react-query';
-import { useState } from 'react';
-
 import { Button } from '@/components/atoms';
 import { css } from '@/styled-system/css';
+
+import { useFollow } from '../../hooks';
 
 type FollowButtonProps = {
   followingId: number;
@@ -15,41 +14,14 @@ export function FollowButton({
   followingId,
   followingInitialValue,
 }: FollowButtonProps) {
-  const [isFollowing, setIsFollowing] = useState(followingInitialValue);
-  const queryClient = useQueryClient();
-
-  const mutation = useMutation({
-    mutationFn: async () => {
-      const response = await fetch('/api/friend', {
-        method: 'PUT',
-        headers: {
-          'Content-Type': 'application/json',
-        },
-        body: JSON.stringify({ followingId: followingId }),
-      });
-      if (!response.ok) {
-        throw new Error('Failed to update following status');
-      }
-      return response.json();
-    },
-    onSuccess: async () => {
-      setIsFollowing(!isFollowing);
-      await queryClient.invalidateQueries({
-        queryKey: ['profileData'],
-      });
-    },
-    onError: (error) => {
-      console.error('Error updating following status:', error);
-    },
+  const { isFollowing, handleFollowing } = useFollow({
+    followingId,
+    followingInitialValue,
   });
 
   const buttonText = isFollowing ? '팔로잉' : '팔로우';
   const buttonVariant = isFollowing ? 'outlined' : 'solid';
   const buttonType = isFollowing ? 'assistive' : 'primary';
-
-  const handleFollowing = () => {
-    mutation.mutate();
-  };
 
   return (
     <Button
