+++ features/profile/hooks/index.ts
@@ -1,3 +1,4 @@
+export * from './use-follow';
 export * from './use-profile-data';
 export * from './use-profile-edit-form';
 export * from './use-profile-image-bottom-sheet';
