+++ features/profile/hooks/use-following-data.tsx
@@ -4,13 +4,13 @@ import { FollowingProps } from '@/features/profile';
 
 import { fetchFollowingData } from '../apis/fetch-following-data';
 
-export const useFollowingData = (id: number) => {
+export const useFollowingData = (id: number, isMyProfile?: boolean) => {
   return useQuery<FollowingProps['data']>({
     queryKey: ['followingStatus', id],
     queryFn: () =>
       fetchFollowingData(id).then((data) => {
         return data.data;
       }),
-    enabled: !!id,
+    enabled: isMyProfile !== undefined && !isMyProfile,
   });
 };
